EDIT : 회원정보 조회 api 수정 및 HomePage.tsx 내 useEffect() 수정
